src/plugins/word_cloud/requirements.txt
pillow
sqlalchemy
spacy
jieba
wordcloud
numpy
nonebot2
nonebot-adapter-onebot

# spacy download en_core_web_sm